Fix regression in package.yml metadata.owner key
As far as I can tell, this key was not intentionally removed and its
presence is missed. Based on issues reported, We're restoring support
rather than fix all the other things that depend on it.

Tolerate having both keys set as long as they are the same.
Fail when metadata.owner and owner are different.

Since we don't know what other tooling reads this owner key, it seems
better to error on ambiguity. It's possible other tooling reads this
value from the package.yml. We can adjust if we find problems.

Related to: rubyatscale/code_ownership#141